Move optimizer driver to Compiler/Driver
