+++ dawn/src/dawn/Compiler/Driver.cpp
@@ -1,0 +1,50 @@
+//===--------------------------------------------------------------------------------*- C++ -*-===//
+//                          _
+//                         | |
+//                       __| | __ ___      ___ ___
+//                      / _` |/ _` \ \ /\ / / '_  |
+//                     | (_| | (_| |\ V  V /| | | |
+//                      \__,_|\__,_| \_/\_/ |_| |_| - Compiler Toolchain
+//
+//
+//  This file is distributed under the MIT License (MIT).
+//  See LICENSE.txt for details.
+//
+//===------------------------------------------------------------------------------------------===//
+
+#include "dawn/Compiler/Driver.h"
+#include "dawn/Compiler/DawnCompiler.h"
+#include "dawn/Compiler/Options.h"
+
+namespace dawn {
+
+std::map<std::string, std::shared_ptr<iir::StencilInstantiation>>
+run(const std::shared_ptr<SIR>& stencilIR, const std::list<PassGroup>& groups,
+    const OptimizerOptions& options) {
+  // Put all options there
+  dawn::Options dawnOptions;
+  // Copy over options passed in
+#define OPT(TYPE, NAME, DEFAULT_VALUE, OPTION, OPTION_SHORT, HELP, VALUE_NAME, HAS_VALUE, F_GROUP) \
+  dawnOptions.NAME = options.NAME;
+#include "dawn/Optimizer/Options.inc"
+#undef OPT
+  DawnCompiler compiler(dawnOptions);
+  return compiler.optimize(compiler.lowerToIIR(stencilIR), groups);
+}
+
+std::map<std::string, std::shared_ptr<iir::StencilInstantiation>>
+run(const std::map<std::string, std::shared_ptr<iir::StencilInstantiation>>&
+        stencilInstantiationMap,
+    const std::list<PassGroup>& groups, const OptimizerOptions& options) {
+  // Put all options there
+  dawn::Options dawnOptions;
+  // Copy over options passed in
+#define OPT(TYPE, NAME, DEFAULT_VALUE, OPTION, OPTION_SHORT, HELP, VALUE_NAME, HAS_VALUE, F_GROUP) \
+  dawnOptions.NAME = options.NAME;
+#include "dawn/Optimizer/Options.inc"
+#undef OPT
+  DawnCompiler compiler(dawnOptions);
+  return compiler.optimize(stencilInstantiationMap, groups);
+}
+
+} // namespace dawn
